Made many changes including:
Changed getComputerChoice to now use switch depending on result from 0-2.
Function playRound, made case insensitive by adding .toLowerCase to both playerSelection & computerSelection.
Made a series of if, else if, and else to declare winner for game depending on outcome.
Next objective will be to figure out how to display the winner of playRound function in console.